+++ packages/astro-components-stencil/src/components/rux-checkbox/rux-checkbox.tsx
@@ -7,6 +7,7 @@ import {
     Element,
     Watch,
 } from '@stencil/core'
+
 import { FormFieldInterface } from '../../common/interfaces.module'
 import { renderHiddenInput } from '../../utils/utils'
 
@@ -30,11 +31,6 @@ export class RuxCheckbox implements FormFieldInterface {
      * The help or explanation text
      */
     @Prop({ attribute: 'help-text' }) helpText?: string
-
-    /**
-     * The validation error text
-     */
-    @Prop({ attribute: 'error-text' }) errorText?: string
 
     /**
      * The checkbox name
@@ -76,11 +72,6 @@ export class RuxCheckbox implements FormFieldInterface {
      * Disables the checkbox via HTML disabled attribute. Checkbox takes on a distinct visual state. Cursor uses the not-allowed system replacement and all keyboard and mouse events are ignored.
      */
     @Prop({ reflect: true }) disabled: boolean = false
-
-    /**
-     * Sets the input as required
-     */
-    @Prop() required: boolean = false
 
     /**
      * Fired when the value of the input changes - [HTMLElement/input_event](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/input_event)
@@ -134,10 +125,8 @@ export class RuxCheckbox implements FormFieldInterface {
             checkboxId,
             checked,
             disabled,
-            errorText,
             helpText,
             name,
-            required,
             value,
             indeterminate,
         } = this
@@ -159,9 +148,7 @@ export class RuxCheckbox implements FormFieldInterface {
                     class={{
                         'rux-checkbox': true,
                         'rux-checkbox--indeterminate': indeterminate,
-                        'rux-checkbox--has-error': required,
-                        'rux-checkbox--has-text':
-                            errorText !== undefined || helpText !== undefined,
+                        'rux-checkbox--has-text': helpText !== undefined,
                     }}
                 >
                     <input
@@ -169,7 +156,6 @@ export class RuxCheckbox implements FormFieldInterface {
                         name={name}
                         id={checkboxId}
                         disabled={disabled}
-                        required={required}
                         checked={checked}
                         //Allows storybook's indetermiante control to take effect.
                         indeterminate={indeterminate}
@@ -190,13 +176,7 @@ export class RuxCheckbox implements FormFieldInterface {
                         </span>
                     </label>
                 </div>
-                {this.helpText && !this.errorText && (
-                    <div class="rux-help-text">{helpText}</div>
-                )}
-
-                {this.errorText && (
-                    <div class="rux-error-text">{errorText}</div>
-                )}
+                {this.helpText && <div class="rux-help-text">{helpText}</div>}
             </div>
         )
     }