+++ packages/astro-components-stencil/src/components/rux-checkbox/test/rux-checkbox.e2e.js
@@ -1,6 +1,6 @@
 describe('Checkbox', () => {
     beforeEach(() => {
-        cy.visitStory('components-checkbox--default-story')
+        cy.visitStory('forms-checkbox--default-story')
     })
     it('renders', () => {
         cy.get('rux-checkbox').should('have.class', 'hydrated')