+++ packages/astro-components-stencil/src/components/rux-input-field/test/rux-input-field.e2e.js
@@ -1,6 +1,6 @@
 describe('Input Field', () => {
     beforeEach(() => {
-        cy.visitStory('components-input-field--default-story')
+        cy.visitStory('forms-input-field--default-story')
     })
     it('renders', () => {
         cy.get('rux-input-field').should('have.class', 'hydrated')