+++ packages/astro-components-stencil/src/components/rux-radio/test/rux-radio.e2e.js
@@ -1,6 +1,6 @@
 describe('Radio', () => {
     beforeEach(() => {
-        cy.visitStory('components-radio--default-story')
+        cy.visitStory('forms-radio--default-story')
     })
     it('renders', () => {
         cy.get('rux-radio').should('have.class', 'hydrated')