+++ packages/astro-components-stencil/src/components/rux-slider/test/rux-slider.e2e.js
@@ -2,7 +2,7 @@
 
 describe('rux-slider', () => {
     beforeEach(() => {
-        cy.visitStory('components-slider--default-story')
+        cy.visitStory('forms-slider--default-story')
     })
     it('renders', () => {
         cy.get('rux-slider').should('have.class', 'hydrated')