+++ packages/astro-components-stencil/src/components/rux-switch/test/rux-switch.spec.tsx
@@ -29,7 +29,6 @@ describe('rux-switch', () => {
             disabled: false,
             errorText: undefined,
             name: '',
-            required: false,
             value: '',
         })
     })