+++ packages/astro-components-stencil/src/components/rux-textarea/test/rux-textarea.e2e.js
@@ -1,6 +1,6 @@
 describe('Textarea', () => {
     beforeEach(() => {
-        cy.visitStory('components-textarea--default-story')
+        cy.visitStory('forms-textarea--default-story')
     })
     it('renders', () => {
         cy.get('rux-textarea').should('have.class', 'hydrated')